refactors Annyang into separate component
