+++ client/components/Annyang.js
@@ -1,23 +1,17 @@
+/* eslint-disable no-alert */
 import React from 'react'
 import annyang from 'annyang'
-import {keywordCheck} from '../utilFunctions'
-import {getKeyWords} from '../store/userStats'
-import {connect} from 'react-redux'
-import Prompt from './Prompt'
-import Help from './Help'
 
 export class Annyang extends React.Component {
   constructor() {
     super()
     this.state = {
-      wordsGot: [],
       said: '',
       buttonClass: 'notRec'
     }
     this.handleChange = this.handleChange.bind(this)
     this.annyangStart = this.annyangStart.bind(this)
     this.annyangStop = this.annyangStop.bind(this)
-    this.handleSubmit = this.handleSubmit.bind(this)
   }
   componentDidMount() {
     if (!annyang) {
@@ -25,12 +19,6 @@ export class Annyang extends React.Component {
         'Speech recognition is not yet working on some browsers. Try using the site in Chrome or you can type your answers.'
       )
     } else {
-      // Let's define a command.
-      var commands = {}
-
-      // Add our commands to annyang
-      annyang.addCommands(commands)
-
       // Start listening.
 
       annyang.addCallback('result', phrases => {
@@ -41,20 +29,7 @@ export class Annyang extends React.Component {
   handleChange() {
     this.setState({said: event.target.value})
   }
-  handleSubmit(event) {
-    event.preventDefault()
 
-    let [wordsGot, wordsNotGot] = keywordCheck(
-      this.state.said,
-      this.props.challenge.keywords
-    )
-
-    this.props.getKeyWords({
-      gotKeywords: wordsGot,
-      notGotKeywords: wordsNotGot
-    })
-    this.props.history.push('/code')
-  }
   annyangStart() {
     this.setState({
       buttonClass: 'Rec'
@@ -69,55 +44,43 @@ export class Annyang extends React.Component {
   }
   render() {
     return (
-      <div className="largeViewBorderCard">
-        <Prompt />
-        <div className="largeViewCard" id="promptAnnyang">
-          <p>
-            Now say how you'd solve the problem! You can record your approach or
-            type it, and you can edit your approach before hitting submit.
-          </p>
-          <div id="rec-stopDiv">
-            <div id="recBtn">
-              <button
-                className={this.state.buttonClass}
-                type="button"
-                onClick={this.annyangStart}
-              />
-              <p>Rec.</p>
-            </div>
-            <div id="stopBtn">
-              <button id="pauseBtn" type="button" onClick={this.annyangStop} />
-              <p>Stop</p>
-            </div>
-          </div>
-
-          <form onSubmit={this.handleSubmit} className="form">
-            <label>You said:</label>
-
-            <textarea
-              className="textBox"
-              rows="5"
-              type="text"
-              name="said"
-              value={this.state.said}
-              onChange={this.handleChange}
-              id="Annyang"
+      <div>
+        <div id="rec-stopDiv">
+          <div id="recBtn">
+            <button
+              className={this.state.buttonClass}
+              type="button"
+              onClick={this.annyangStart}
             />
-
-            <i className="far fa-comments" />
-            <input id="submitBtn" type="submit" />
-          </form>
+            <p>Rec.</p>
+          </div>
+          <div id="stopBtn">
+            <button id="pauseBtn" type="button" onClick={this.annyangStop} />
+            <p>Stop</p>
+          </div>
         </div>
-        <Help />
+        <form
+          onSubmit={() => this.props.handleSubmit(this.state.said)}
+          className="form"
+        >
+          <label>You said:</label>
+
+          <textarea
+            className="textBox"
+            rows="5"
+            type="text"
+            name="said"
+            value={this.state.said}
+            onChange={this.handleChange}
+            id="Annyang"
+          />
+
+          <i className="far fa-comments" />
+          <input id="submitBtn" type="submit" />
+        </form>
       </div>
     )
   }
 }
 
-const mapState = state => ({
-  challenge: state.problems.selected
-})
-
-const mapDispatch = {getKeyWords}
-
-export default connect(mapState, mapDispatch)(Annyang)
+export default Annyang
