+++ client/routes.js
@@ -7,7 +7,7 @@ import {
   Signup,
   UserHome,
   GuestHome,
-  Annyang,
+  Approach,
   UserStats,
   Code,
   Solutions,
@@ -27,6 +27,7 @@ import {
 import Examples from './components/Examples'
 import {me} from './store'
 import {ChallengesView} from './components/ChallengesView'
+
 //import {InterviewBot} from './components/interviewBot'
 
 /**
@@ -46,7 +47,7 @@ class Routes extends Component {
         <Route path="/login" component={GuestHome} />
         <Route path="/signup" component={GuestHome} />
         <Route path="/prompt" component={Read} />
-        <Route path="/approach" component={Annyang} />
+        <Route path="/approach" component={Approach} />
         <Route path="/examples" component={Examples} />
         <Route path="/stats" component={UserStats} />
         <Route path="/code" component={Code} />
